requirements.txt
# Core Trading Libraries
MetaTrader5==5.0.45
ccxt==4.1.22  # Para comparación/backup con otros exchanges

# Data Processing and Analysis
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
statsmodels==0.14.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.15.0
torch==2.0.1
optuna==3.5.0  # Para optimización de hiperparámetros
shap==0.44.0  # Para interpretabilidad de modelos

# Technical Analysis
ta-lib==0.4.28  # Requiere instalación separada de TA-Lib binaries
pandas-ta==0.3.14b0
tulipy==0.4.0  # Alternativa ligera a TA-Lib

# Visualization and Dashboard
plotly==5.18.0
streamlit==1.29.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.0

# Database and Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.0.1  # Para caché y mensajería
pymongo==4.6.1  # MongoDB si se usa

# API and Web
fastapi==0.104.1
uvicorn==0.25.0
requests==2.31.0
websocket-client==1.7.0
aiohttp==3.9.1

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and Monitoring
loguru==0.7.2
sentry-sdk==1.39.1  # Para error tracking
prometheus-client==0.19.0  # Para métricas

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0  # Para generar datos de prueba

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3
mkdocs-material==9.5.2

# Utilities
schedule==1.2.0  # Para tareas programadas
python-dateutil==2.8.2
pytz==2023.3.post1
humanize==4.9.0
tqdm==4.66.1  # Progress bars
colorama==0.4.6  # Colored terminal output
tabulate==0.9.0  # Pretty tables

# Notifications
python-telegram-bot==20.7
discord.py==2.3.2
slack-sdk==3.26.1
sendgrid==6.11.0  # Email notifications

# Financial Data
yfinance==0.2.33  # Para datos históricos
alpha-vantage==2.3.1
finnhub-python==2.4.19

# Risk and Portfolio Management
cvxpy==1.4.1  # Para optimización convexa
empyrical==0.5.5  # Métricas de performance
pyfolio==0.9.2  # Análisis de portfolio
riskfolio-lib==4.5.1  # Optimización de portfolio avanzada

# Backtesting
backtrader==1.9.78.123
zipline-reloaded==3.0.3
vectorbt==0.26.1  # Backtesting vectorizado rápido

# Time Series Analysis
arch==6.2.0  # Modelos GARCH
pmdarima==2.0.4  # Auto ARIMA
prophet==1.1.5  # Facebook Prophet para forecasting
statsforecast==1.6.0  # Statistical forecasting

# Natural Language Processing (para análisis de noticias)
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2  # Análisis de sentimiento

# Parallel Processing
joblib==1.3.2
multiprocess==0.70.15
dask==2023.12.1  # Para procesamiento distribuido

# Code Quality
black==23.12.1  # Formatter
flake8==7.0.0  # Linter
mypy==1.8.0  # Type checking
isort==5.13.2  # Import sorting
pre-commit==3.6.0  # Git hooks

# Security
cryptography==41.0.7
python-jose==3.3.0  # JWT tokens
passlib==1.7.4  # Password hashing

# Development Tools
ipython==8.19.0
jupyter==1.0.0
notebook==7.0.6
jupyterlab==4.0.9

# Deployment
gunicorn==21.2.0
supervisor==4.2.5
docker==7.0.0
kubernetes==29.0.0

# Profiling and Optimization
memory-profiler==0.61.0
line-profiler==4.1.1
py-spy==0.3.14  # Sampling profiler

# Additional ML Libraries
catboost==1.2.2
imbalanced-learn==0.11.0  # Para datos desbalanceados
feature-engine==1.6.2  # Feature engineering
mlflow==2.9.2  # ML lifecycle management

# Report Generation
reportlab==4.0.8  # PDF generation
xlsxwriter==3.1.9  # Excel reports
jinja2==3.1.2  # Templates
weasyprint==60.2  # HTML to PDF
pdfkit==1.0.0  # wkhtmltopdf wrapper

# System Monitoring
psutil==5.9.7  # System metrics
gputil==1.4.0  # GPU monitoring si se usa


# ============================================
# NOTAS DE INSTALACIÓN
# ============================================

# 1. TA-Lib requiere instalación de binarios del sistema:
#    - Windows: descargar wheel de https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
#    - Linux: sudo apt-get install ta-lib
#    - MacOS: brew install ta-lib

# 2. Para GPU support en TensorFlow:
#    pip install tensorflow[and-cuda]

# 3. MetaTrader5 solo funciona en Windows
#    Para desarrollo en Linux/Mac usar: pip install MetaTrader5==5.0.45 --only-binary :all:

# 4. Algunas librerías pueden requerir compiladores C++:
#    - Windows: instalar Visual Studio Build Tools
#    - Linux: sudo apt-get install build-essential
#    - MacOS: xcode-select --install

# 5. Para producción, congelar versiones exactas:
#    pip freeze > requirements-prod.txt

# 6. Orden de instalación recomendado:
#    pip install numpy scipy  # Primero
#    pip install -r requirements.txt  # Después

# 7. Environments recomendados:
#    - Desarrollo: Python 3.10.x
#    - Producción: Python 3.10.x (mismo que desarrollo)
#    - Usar virtual environment: python -m venv venv

# 8. Para instalar todo:
#    pip install -r requirements.txt

# 9. Para desarrollo instalar también:
#    pip install -r requirements-dev.txt